feat: use start regtest lib
